% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorspace.functions.R
\name{angv}
\alias{angv}
\title{Angle between two vectors}
\usage{
angv(vec1, vec2, type = "deg")
}
\arguments{
\item{vec1}{vector 1}

\item{vec2}{vector 2}
}
\value{
The function will XX
}
\description{
Angle between two vectors
}
\details{
Angle between two vectors. Handy for working in the gamut
}
