% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerov.functions.R
\name{approx.normalize}
\alias{approx.normalize}
\title{approx.normalize}
\usage{
approx.normalize(wf.info, include.jacobian = FALSE, plotQ = FALSE)
}
\arguments{
\item{XXXX}{}
}
\value{
A two column matrix. The first column is the x-axis over which the wave function
was obtained. The second column is the normalized wave function solution.
}
\description{
Approximately normalize a 1D-wave function
}
\details{
XXXX
}
\examples{
XXX

}
\references{
Quantum Chemistry by Ira Levine. Any edition.
}
