% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyatomic.harmonic.vibs.functions.R
\name{atom.colors}
\alias{atom.colors}
\title{XXXX}
\usage{
atom.colors(atoms.char.vec)
}
\arguments{
\item{nqn}{XXXX}
}
\value{
XXXX
}
\description{
XXXX
}
\details{
XXXX
}
\examples{
library(che302r)
XXXX
}
\references{
XXXXX
}
