% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyatomic.harmonic.vibs.functions.R
\name{atom2color}
\alias{atom2color}
\title{XXXX}
\usage{
atom2color(atom.char)
}
\arguments{
\item{nqn}{XXXX}
}
\value{
XXXXX
}
\description{
XXXX
}
\details{
XXXX
}
\examples{
library(che302r)
XXXX
}
\references{
XXXXX
}
