% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorspace.functions.R
\name{color_partition}
\alias{color_partition}
\title{color_partition}
\usage{
color_partition(wavelength)
}
\arguments{
\item{wavelength}{a wavelength in nm}
}
\value{
A hexcode for an RGB color
}
\description{
This is a function takes wavelength in nm and returns simple color name
}
\details{
This is a function takes wavelength in nm and returns simple color name.
It works the same way as wavelengthToColor
}
\examples{
wavelengthToColor(536)

}
