% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorspace.functions.R
\name{color_purity}
\alias{color_purity}
\title{Compute "purity" of a point in xy-chromaticity space}
\usage{
color_purity(xc, yc, xw = 0.31, yw = 0.33, percentQ = T, printQ = F)
}
\arguments{
\item{xc}{x-chromaticity}

\item{yc}{y-chromaticity}

\item{xw}{x-white point. Default is D65 xw=0.31}

\item{yw}{y-white point. Default is D65 yw=0.33}

\item{percentQ}{Compute as a percent?}

\item{printQ}{Print more info?}
}
\value{
A chromaticity's purity with respect to the white point
}
\description{
Compute "purity" of a point in xy-chromaticity space
}
\details{
Compute "purity" of a point in xy-chromaticity space  (i.e. "closeness)
to the locus convex hull wrt/ a chosen whit point). Assumes as default D65 white point.
}
