% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorspace.functions.R
\name{compute_hue}
\alias{compute_hue}
\title{Compute hue}
\usage{
compute_hue(xc, yc, xw = 0.31, yw = 0.33)
}
\arguments{
\item{xc}{x-chromaticity}

\item{yc}{y-chromaticity}

\item{xw}{x-white point. Default is D65 xw=0.31}

\item{yw}{y-white point. Default is D65 yw=0.33}
}
\value{
A chromaticity's purity with respect to the white point
}
\description{
Compute hue
}
\details{
Compute the approximate closest point on the convex hull of the gamut to
chromaticity coords xc, yc wrt white point xw, yw. D65 white point is the
default.
}
