% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_slit_functions.R
\name{dsi}
\alias{dsi}
\title{Double slit intensity function}
\usage{
dsi(theta, d.slit.sep, a.slit.width, lambda, Imax = 1)
}
\arguments{
\item{theta}{direction angle between -pi/2 and pi/2 (-90 deg, 90deg)}

\item{d.slit.sep}{slit separation}

\item{a.slit.width}{slit width}

\item{lambda}{wavelength}

\item{I.max}{largest intensity to use}
}
\value{
Double slit intensity.
}
\description{
This is a function to evaluate the double slit intensity function
}
\details{
This is a function to evaluate the double slit intensity function.
}
\examples{
library(che302r)

dsi(pi/2, 10, 0.1, 650)

#theta <- seq(from=-pi/2, to=pi/2, length.out=1000)
theta <- seq(from=-pi/4, to=pi/4, length.out=1000)
lam <- 650
d <- 6*lam
a <- 2*lam
I <- dsi(theta, d.slit.sep = d, a.slit.width = a, lambda = lam)
plot(theta*180/pi, I, typ="l", xlab="theta (deg)")

}
