% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorspace.functions.R
\name{ed}
\alias{ed}
\title{Euclidean distance}
\usage{
ed(vec1, vec2)
}
\arguments{
\item{vec1}{a vector}

\item{vec2}{a vector}
}
\value{
The function will XX
}
\description{
Euclidean distance in a plane
}
\details{
Euclidean distance in a plane
}
