% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_script}
\alias{get_script}
\title{Function to download an example script}
\usage{
get_script(script_name, url_head = NULL, download_location = NULL)
}
\arguments{
\item{script_name}{Name of the script}

\item{url_head}{Location of the script (optional). If NULL, default location is used.}

\item{download_location}{Where to download the script. If NULL, default location is used.}
}
\value{
The script should pop up.
}
\description{
The function will download script that is an example
}
\details{
The function will download an example script. Default url_head is currently:

https://raw.githubusercontent.com/npetraco/che302r/refs/heads/main/inst/scripts_bank.

Default download_location is currently: tempdir().
}
\examples{
get_script("XXXXXX.R")

}
