% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{logsumexp}
\alias{logsumexp}
\title{Log sum exp trick}
\usage{
logsumexp(logv)
}
\arguments{
\item{logv}{vector of log-ed quantities to add up}
}
\value{
The function will XX
}
\description{
Log sum exp trick to prevent over/underflow
}
\details{
Handy for calculating logZ from a vector of log Boltzmann factors.
}
