% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorspace.functions.R
\name{magv}
\alias{magv}
\title{Magnitude of a vector}
\usage{
magv(vec1)
}
\arguments{
\item{vec1}{a vector}

\item{vec2}{}
}
\value{
The function will XX
}
\description{
Magnitude of a vector
}
\details{
Magnitude of a vector. Handy for working in the gamut
}
