% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorspace.functions.R
\name{make.alychne}
\alias{make.alychne}
\title{Generate alychne data or line}
\usage{
make.alychne(
  lambda.left = 380,
  lambda.right = 780,
  type = "data",
  num.pts = 1000
)
}
\arguments{
\item{lambda.left}{XX}

\item{lambda.right}{XX}

\item{type}{XX}

\item{num.pts}{XX}
}
\value{
The function will XX
}
\description{
XX
}
\details{
XX
}
