% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorspace.functions.R
\name{make.spectral.locus}
\alias{make.spectral.locus}
\title{Generate spectral locus data}
\usage{
make.spectral.locus(
  delta.lambda = 1,
  num.lambda = NULL,
  lambda.start = 380,
  lambda.stop = 780
)
}
\arguments{
\item{xch}{XX}

\item{ych}{XX}

\item{lambda}{XX}
}
\value{
The function will XX
}
\description{
XX
}
\details{
XX
}
