% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.mech.functions.R
\name{multinom}
\alias{multinom}
\title{multinomial coefficient}
\usage{
multinom(x, logQ = F)
}
\arguments{
\item{x}{a macrostate, i.e. a vector of one-particle state occupancies}

\item{logQ}{whether or not to return the log of the multinomial coefficient. Default is FALSE}
}
\value{
multinomial coefficient.
}
\description{
This function computes the multinomial coefficient
}
\details{
This is a function computes the multinomial coefficient.
}
\examples{
library(che302r)

x <- c(3,2,8,0,0,1,0,0)
N <- sum(x)
factorial(N)/prod(factorial(x))
multinom(x)

multinom(c(3,3,7,0,0,1,0,0))

multinom(c(2,4,1,2,2,0,1,2), logQ = T)

}
