% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyatomic.harmonic.vibs.functions.R
\name{oscillation.f}
\alias{oscillation.f}
\title{XXXX}
\usage{
oscillation.f(xyz0.coords, Qj.mat, t.coord)
}
\arguments{
\item{nqn}{XXXX}
}
\value{
XXXX
}
\description{
XXXX
}
\details{
XXXX
}
\examples{
library(che302r)
XXXX
}
\references{
XXXXX
}
