% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.mech.functions.R
\name{piab.energy}
\alias{piab.energy}
\title{Particle in a box energy}
\usage{
piab.energy(n, m, L)
}
\arguments{
\item{n}{vector of PIAB n quantum numbers: 1, 2, 3, .....}

\item{m}{particle mass in kg}

\item{L}{vector of box side lengths in meters}
}
\value{
a PIAB energy in joules.
}
\description{
This function computes the particle in a box energy
}
\details{
This is a function computes the particle in a box energy. If n is a single positive integer
and L is a positive real number, the function computed the 1D particle in a box energy.
}
\examples{
library(che302r)

# 1D PIAB example:
num.states <- 4        # Number of PIAB states
L.box      <- 50e-6    # Box length
m.particle <- 1.78e-37 # Mass of particle(s)
TT         <- 77       # Temperature
N          <- 2.3*N.A  # Total number of particles

# PIAB energies:
e <- piab.energy(1:num.states, m = m.particle, L = L.box)
e

}
