% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planck.distribution.functions.R
\name{planck.distribution.lambda}
\alias{planck.distribution.lambda}
\title{Plank distribution as a function of wavelength.}
\usage{
planck.distribution.lambda(lam.min, lam.max, Temp, plotQ = FALSE)
}
\arguments{
\item{lam.min}{The smallest wavelength to use. Use something at least little larger than 0.}

\item{lam.max}{The largest wavelength to use.}

\item{Temp}{The temperature in Kelvin}

\item{plotQ}{Whether or not to plot the distribution}
}
\value{
A two column matrix. The first column is the lambda-axis over which the Planck
distribution was computed. The second column is the Planck distribution values
}
\description{
This is a function to plot the Plank distribution as a function of wavelength.
}
\details{
This is a function to plot the Plank distribution as a function of wavelength.
}
\examples{
#Now USE the function planck.distribution.lambda below:
planck.distribution.lambda(10e-9,10000e-9,1500)
planck.distribution.lambda(10e-9,12000e-9,1500)

}
