% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planck.distribution.functions.R
\name{planck.distribution.nu}
\alias{planck.distribution.nu}
\title{Plank distribution as a function of frequency.}
\usage{
planck.distribution.nu(nu.min, nu.max, Temp, plotQ = FALSE)
}
\arguments{
\item{Temp}{The temperature in Kelvin}

\item{plotQ}{Whether or not to plot the distribution}

\item{nut.min}{The smallest frequency to use. Use something at least little larger than 0.}

\item{nut.max}{The largest frequency to use.}
}
\value{
A two column matrix. The first column is the nu-axis over which the Planck
distribution was computed. The second column is the Planck distribution values
}
\description{
This is a function to plot the Plank distribution as a function of frequency.
}
\details{
This is a function to plot the Plank distribution as a function of frequency.
}
\examples{
#USE the function planck.distribution.nu below:
planck.distribution.nu(10e10, 6e14, 1500, plotQ = T)

}
