% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planck.distribution.functions.R
\name{planck.distribution}
\alias{planck.distribution}
\title{Plank distribution as a function of wavenumber, wavelength or frequency.}
\usage{
planck.distribution(typ, x.min, x.max, Temp, plotQ = FALSE)
}
\arguments{
\item{typ}{Scale units for witch to compute the Planck distribution. Choices are: "wavenumber", "wavelength", "frequency"}

\item{x.min}{The smallest domain value to use. Use something at least little larger than 0}

\item{x.max}{The largest domain to use}

\item{Temp}{The temperature in Kelvin}

\item{plotQ}{Whether or not to plot the distribution}
}
\value{
A two column matrix. The first column is the x-axis over which the Planck
distribution was computed (m^-1, m or s^-1). The second column is the Planck distribution values
}
\description{
This is a function to plot the Plank distribution as a function of wavenumber, wavelength or frequency.
}
\details{
This wrapper is a function to plot the Plank distribution as a function of wavenumber, wavelength or frequency.
}
\examples{
#USE the function planck.distribution below:
planck.distribution(typ = "wavenumber", 10000, 3000000,  1500, plotQ = T)
planck.distribution(typ = "wavelength", 10e-9, 10000e-9, 1500, plotQ = T)
planck.distribution(typ = "frequency",  10e10, 6e14,     1500, plotQ = T)

}
