% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorspace.functions.R
\name{plot_spectral_locus}
\alias{plot_spectral_locus}
\title{Plot the outline of the gamut with and without partitions}
\usage{
plot_spectral_locus(Y.level, white.point = c(0.31, 0.33), type = "sRGB")
}
\arguments{
\item{Y.level}{The brightness value (Y) to use}

\item{white.point}{White point. Default is D65}

\item{type}{Color sub-gamut to use, Adobe, sRGB, P3, rough.hue, wl2c}

\item{partitionsQ}{Whether or not to show partitions between the rough hues}
}
\value{
The function will XX
}
\description{
Plot the outline of the gamut with and without partitions
}
\details{
Plot the outline of the gamut with and without partitions
}
