% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorspace.functions.R
\name{prjv}
\alias{prjv}
\title{Scalar projection of vector 1 on to vector 2}
\usage{
prjv(vec1, vec2)
}
\arguments{
\item{vec1}{vector 1}

\item{vec2}{vector 2}
}
\value{
The function will XX
}
\description{
Scalar projection of vector 1 on to vector 2
}
\details{
Scalar projection of vector 1 on to vector 2. Usually vector 1 is the
shorter of the two vectors, and vector 2 is the longer.
Handy for working in the gamut.
}
