% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorspace.functions.R
\name{purpleQ}
\alias{purpleQ}
\title{Test to see if a point in the gamut is in the purple region}
\usage{
purpleQ(
  cp,
  white.point = c(0.31, 0.33),
  left.purple.vertex = c(0.175560232, 0.005293837),
  right.purple.vertex = c(0.73469, 0.26531),
  tol = 1e-05
)
}
\arguments{
\item{cp}{Chromaticity point in the gamut}

\item{white.point}{A white point. Default is D65}

\item{left.purple.vertex}{A left purple vertex. Default is the one in ciexyz1931}

\item{right.purple.vertex}{A right purple vertex. Default is the one in ciexyz1931}
}
\value{
The function will XX
}
\description{
Test to see if a point in the gamut is in the purple region
}
\details{
Test to see if a point in the gamut is in the purple region. The purple
region is a convex hull.
}
