% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_slit_functions.R
\name{ssi}
\alias{ssi}
\title{Single slit intensity function}
\usage{
ssi(theta, a.slit.width, lambda, Imax = 1)
}
\arguments{
\item{theta}{direction angle between -pi/2 and pi/2 (-90 deg, 90deg)}

\item{a.slit.width}{slit width}

\item{lambda}{wavelength}

\item{I.max}{largest intensity to use}
}
\value{
Single slit intensity.
}
\description{
This is a function to evaluate the single slit intensity function
}
\details{
This is a function to evaluate the single slit intensity function. It isn't really necessary
because we can just make the slit separation (d.slit.sep) 0 in dsi() to get ssi(). It's here for
convenience.
}
\examples{
library(che302r)

ssi(pi/2, 0.1, 650)

theta <- seq(from=-pi/2, to=pi/2, length.out=1000)
#theta <- seq(from=-pi/4, to=pi/4, length.out=1000)
lam   <- 650
a     <- 3*lam
I     <- ssi(theta, a.slit.width = a, lambda = lam)
plot(theta*180/pi, I, typ="l", xlab="theta (deg)")

}
