% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorspace.functions.R
\name{tria}
\alias{tria}
\title{Area for a triangle by 3x3 determinant of vertices}
\usage{
tria(vtx1, vtx2, vtx3)
}
\arguments{
\item{vtx1}{vertex 1}

\item{vtx2}{vertex 2}

\item{vtx3}{vertex 3}
}
\value{
The function will XX
}
\description{
Area for a triangle by 3x3 determinant of vertices
}
\details{
Area for a triangle by 3x3 determinant of vertices.
Handy for testing if chromaticity point is in purple region.
}
