% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{update_che302r}
\alias{update_che302r}
\title{Easily update the che302r library by installing the current version from the github site}
\usage{
update_che302r()
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Easily update the che302r library by installing the current version from the github site
}
