% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerov.functions.R
\name{V}
\alias{V}
\title{Potential energy function for use with Numerov's procedure}
\usage{
V(xax, potential.name, l = NULL)
}
\arguments{
\item{xax}{The x-axis over which to find a solution to the Schrodinger equation with the requested potential}

\item{potential.name}{Name of the potential energy function to insert into the Schrodinger equation. Choices are: "box", "harmonic", "anharmonic", "radial"}

\item{l}{The l quantum number if the "radial" potential energy function is requested}
}
\value{
Values of the potential along the specified x-axis.
}
\description{
Potential energy function
}
\details{
Potential energy functions for use with Numerov's procedure to solve Schrodinger
equations implemented in this package. If Laguerre solutions are desired for orbital
calculations, choose the "radial" potential energy function and specify an l quantum number.
}
\examples{
# Potential energy function for particle in a box:
# Domain (x-axis):
dx    <- 0.01                            # Increment on x-axis
x.min <- (0)                             # Start of x-axis. Well into (left) classically forbidden region
x.max <- 1.5                             # End of x-axis. Well into the (right) classically forbidden region
x     <- seq(from=x.min,to=x.max,by=dx)  # x-axis

potential<-"box"
plot(x,V(x,potential))

# Potential energy function for anharmonic oscillator:
#Domain (x-axis):
dx    <- 0.01
x.min <- (-0.8)
x.max <- 5
x     <- seq(from=x.min,to=x.max,by=dx)
plot(x,V(x,"anharmonic"))

}
\references{
Quantum Chemistry by Ira Levine. Any edition.
}
