% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorspace.functions.R
\name{wavelengthToColor}
\alias{wavelengthToColor}
\title{wavelengthToColor.}
\usage{
wavelengthToColor(wavelength)
}
\arguments{
\item{wavelength}{a wavelength in nm}
}
\value{
A hexcode for an RGB color
}
\description{
This is a function takes wavelength in nm and returns an rgb-alpha value in hex.
}
\details{
This is a function takes wavelength in nm and returns an rgb-alpha value in hex.
From: http://www.physics.sfasu.edu/astro/color/spectra.html converted to R. I think this
guy may have written the original code however: https://earlglynn.github.io/
}
\examples{
wavelengthToColor(536)

}
