% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorspace.functions.R
\name{xyl.lut}
\alias{xyl.lut}
\title{xy-chromaticity and lambda lookup table}
\usage{
xyl.lut(xch, ych, lambda = NULL, delta.lambda = 0.1, tol = 0.005, printQ = F)
}
\arguments{
\item{xch}{XX}

\item{ych}{XX}

\item{lambda}{XX}
}
\value{
The function will XX
}
\description{
xy-chromaticity and lambda lookup table
}
\details{
xy-chromaticity and lambda lookup table so we can go between xy
normalized chromaticities and corresponding lambdas. Have to do this as a
lookup table because the gamut is not a function and doesn't have an easy
to define inverse.
}
