% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorspace.functions.R
\name{xyY_swatch}
\alias{xyY_swatch}
\title{Plot a color with specified chromaticity and brightness values}
\usage{
xyY_swatch(xyY.vec, type = "Adobe")
}
\arguments{
\item{type}{subgamut to use: Adobe, sRGB or P3}

\item{xyY.vex}{an x, y, Y color specification}
}
\value{
The function will XX
}
\description{
Plot a color with specified chromaticity and brightness values
}
\details{
Plot a color with specified chromaticity and brightness values
}
