% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerov.functions.R
\name{approx.normalize}
\alias{approx.normalize}
\title{approx.normalize}
\usage{
approx.normalize(wf.info, include.jacobian = FALSE, plotQ = FALSE)
}
\arguments{
\item{wf.info}{Output from numerov.procedure}

\item{include.jacobian}{For use with radial wave functions. To normalize these wave functions the Jacobian must be included (TRUE)}

\item{plotQ}{Whether or not to plot the un-normalized probability density. For debugging.}
}
\value{
A two column matrix. The first column is the x-axis over which the wave function
was obtained. The second column is the normalized wave function solution.
}
\description{
Approximately normalize a 1D-wave function
}
\details{
Approximately normalize a 1D-wave function spit out of numerov.procedure.
}
\examples{
cf. examples in numerov.procedure

}
\references{
Quantum Chemistry by Ira Levine. Any edition.
}
