% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{logsumexp}
\alias{logsumexp}
\title{Log sum exp trick}
\usage{
logsumexp(logv)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Handy for calculating Z from a vector of log potentials.
}
