% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{update_che302r}
\alias{update_che302r}
\title{Update che302r package}
\usage{
update_che302r()
}
\value{
Nothing
}
\description{
Easily update the che302r library by installing the current version from the github site.
}
\details{
Easily update the che302r library by installing the current version from the github site
}
